Make Lab pad jump held-level and ship corner force-wave seats.
Controller A was edge-pulsed then cleared in wasm, so jumps raced Bevy/GGRS while touch Space held cleanly; also emit CornerPulse on trampoline hits and keep host/guest seats stable.

Handoff_2026-08-30_arathyll_bifrost_cta_corners

+++ crates/bifrost_client/src/input_focus.rs
@@ -16,7 +16,7 @@ pub struct InputFocus {
     last_mouse_x: f32,
     last_mouse_y: f32,
     /// Sticky mouse move bits — cursor-lost fallback only (not mid-chase thrust).
-    pub mouse_move_sticky: u8,
+    pub mouse_move_sticky: u16,
 }
 
 /// Visual paddle pose used for mouse aim so online rollback / input-delay
